requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
alembic==1.13.0

# AI and LangChain
openai==1.3.0
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.1
langgraph==0.0.15

# Vector Database
chromadb==0.4.17

# Data Processing
numpy==1.24.3
pandas==2.1.3

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
requests==2.31.0
aiofiles==23.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging
structlog==23.2.0